modify datatype from worker_t to int

+++ csrc/core/all2all.cu
@@ -6,7 +6,7 @@ using namespace hetu;
 void HetuGPUTable::all2allExchangeShape(const size_t *shape, size_t *shape_out) {
   assert(shape != shape_out);
   checkCudaErrors(ncclGroupStart());
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       shape + i, 1, ncclUint64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -18,7 +18,7 @@ void HetuGPUTable::all2allExchangeShape(const size_t *shape, size_t *shape_out) 
 void HetuGPUTable::all2allExchangeQuery() {
   checkCudaErrors(ncclGroupStart());
   size_t snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_query_idx_[0] + snd_offset, cur_batch_.u_shape[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -29,7 +29,7 @@ void HetuGPUTable::all2allExchangeQuery() {
   checkCudaErrors(ncclGroupEnd());
   checkCudaErrors(ncclGroupStart());
   snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_query_version_[0] + snd_offset, cur_batch_.u_shape[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -43,7 +43,7 @@ void HetuGPUTable::all2allExchangeQuery() {
   // gradient part, using prev_batch
   checkCudaErrors(ncclGroupStart());
   snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_query_gradient_idx_[0] + snd_offset, prev_batch_.u_shape[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -55,7 +55,7 @@ void HetuGPUTable::all2allExchangeQuery() {
 
   checkCudaErrors(ncclGroupStart());
   snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_query_updates_[0] + snd_offset, prev_batch_.u_shape[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -67,7 +67,7 @@ void HetuGPUTable::all2allExchangeQuery() {
 
   checkCudaErrors(ncclGroupStart());
   snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_query_val_[0] + snd_offset * kEmbeddingWidth, prev_batch_.u_shape[i] * kEmbeddingWidth,
       ncclFloat32, i, communicator_, stream_main_));
@@ -84,7 +84,7 @@ void HetuGPUTable::all2allExchangeQuery() {
 void HetuGPUTable::all2allReturnOutdated() {
   checkCudaErrors(ncclGroupStart());
   size_t snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_return_outdated_[0] + snd_offset, cur_batch_.u_shape_exchanged[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -98,7 +98,7 @@ void HetuGPUTable::all2allReturnOutdated() {
 void HetuGPUTable::all2allReturnValue() {
   checkCudaErrors(ncclGroupStart());
   size_t snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_return_version_[0] + snd_offset, cur_batch_.u_shape[i], ncclInt64, i, communicator_, stream_main_));
     checkCudaErrors(ncclRecv(
@@ -110,7 +110,7 @@ void HetuGPUTable::all2allReturnValue() {
 
   checkCudaErrors(ncclGroupStart());
   snd_offset = 0, rcvd_offset = 0;
-  for (int i = 0; i < (int)nrank_; i++) {
+  for (int i = 0; i < nrank_; i++) {
     checkCudaErrors(ncclSend(
       d_return_val_[0] + snd_offset * kEmbeddingWidth, cur_batch_.u_shape[i] * kEmbeddingWidth,
       ncclFloat32, i, communicator_, stream_main_));
